+++ SimG4Core/Application/src/TrackingAction.cc
@@ -1,5 +1,5 @@
 #include "SimG4Core/Application/interface/TrackingAction.h"
-#include "SimG4Core/Application/interface/CMSG4TrackInterface.h"
+#include "SimG4Core/Physics/interface/CMSG4TrackInterface.h"
 
 #include "SimG4Core/Notification/interface/CurrentG4Track.h"
 #include "SimG4Core/Notification/interface/BeginOfTrack.h"
